fix firebase atuh
