+++ src/atoms/authUser.js
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+
+export const authUserAtom = atom(null);
+
