+++ src/components/CreateComponent/index.jsx
@@ -1,22 +1,15 @@
-import React, { useEffect } from "react";
-import { useNavigate } from "react-router-dom/dist";
-import { useAuth } from "../../hooks/useAuth";
+import React from "react";
 
 import CanvasArea from "./CanvasArea";
 import Sidebar from "./Sidebar";
 
 const CreateComponent = () => {
-    const navigate = useNavigate();
-    const { userAuth } = useAuth();
-    useEffect(() => {
-        if(!userAuth)navigate("/")
-    })
-    return (
-        <div className="flex gap-x-4">
-            <CanvasArea />
-            <Sidebar />
-        </div>
-    );
-}
+  return (
+    <div className="flex gap-x-4">
+      <CanvasArea />
+      <Sidebar />
+    </div>
+  );
+};
 
 export default CreateComponent;
