+++ src/components/RouterConfig.jsx
@@ -1,5 +1,6 @@
 import React, { lazy, Suspense } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Container from "./Container";
 
 export const RouterConfig = () => {
   const CreateMap = lazy(() => import("./CreateMap"));
@@ -13,12 +14,14 @@ export const RouterConfig = () => {
     <>
       <BrowserRouter>
         <Suspense fallback={Loading}>
-          <Routes>
-            <Route path="/" element={<Login />} />
-            <Route path="/createMap" element={<CreateMap />} />
-            <Route path="/createComponent" element={<CreateComponent />} />
-            <Route path="/gallery" element={<Gallery />} />
-          </Routes>
+          <Container>
+            <Routes>
+              <Route path="/" element={<Login />} />
+              <Route path="/createMap" element={<CreateMap />} />
+              <Route path="/createComponent" element={<CreateComponent />} />
+              <Route path="/gallery" element={<Gallery />} />
+            </Routes>
+          </Container>
         </Suspense>
       </BrowserRouter>
     </>
